+++ test/integration/base.py
@@ -1,7 +1,6 @@
-import errno
-import socket
 import unittest
 
+from .pelikan_client import PelikanClient
 from .pelikan_server import PelikanServer
 
 SERVER_PORT = 12345
@@ -16,41 +15,21 @@ class PelikanTest(unittest.TestCase):
             server_port=SERVER_PORT,
             admin_port=ADMIN_PORT
         )
-        self.client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.client.connect(('localhost', SERVER_PORT))
-        self.admin = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.admin.connect(('localhost', ADMIN_PORT))
+        self.client = PelikanClient(
+            server_port=SERVER_PORT,
+            admin_port=ADMIN_PORT
+        )
 
     def tearDown(self):
         self.server.stop()
 
     def assertRead(self, expected):
-        read = self.client.recv(len(expected))
+        read = self.client.read(len(expected))
         self.assertEqual(expected, read)
 
-    def getStats(self):
-        self.admin.sendall('stats\r\n')
-        data = ''
-        self.admin.setblocking(False)
-        while True:
-            try:
-                data += self.admin.recv(1024)
-            except socket.error, e:
-                err = e.args[0]
-                if err == errno.EAGAIN or err == errno.EWOULDBLOCK:
-                    if len(data) == 0:
-                        continue
-                    else:
-                        break
-                else:
-                    raise
-        self.admin.setblocking(True)
-        # this NULL is kinda unexpected for me
-        self.assertEqual(data[-6:], 'END\r\n\0')
-        return dict(line.split(' ')[1:] for line in data[:-6].strip().split('\r\n'))
 
     def assertMetrics(self, *args):
-        stats = self.getStats()
+        stats = self.client.getStats()
         for (k, v) in args:
             self.assertEqual(
                 stats.get(k, None),
