+++ test/integration/test_twemcache_basic.py
@@ -2,6 +2,6 @@ from .base import TwemcacheTest
 
 class TwemcacheBasicTest(TwemcacheTest):
     def test_miss(self):
-        self.client.sendall('get foo\r\n')
+        self.client.write('get foo\r\n')
         self.assertRead('END')
         self.assertMetrics(('request_parse', 1), ('get', 1), ('get_key_miss', 1))
